// src/types/player-api-events.ts
export interface AlbumDetails {
  responseContext: ResponseContext;
  contents: Contents;
  currentVideoEndpoint: CurrentVideoEndpoint;
  trackingParams: string;
  playerOverlays: PlayerOverlays;
  videoReporting: VideoReporting;
}

export interface Contents {
  singleColumnMusicWatchNextResultsRenderer: SingleColumnMusicWatchNextResultsRenderer;
}

export interface SingleColumnMusicWatchNextResultsRenderer {
  tabbedRenderer: TabbedRenderer;
}
// ...
export interface CurrentVideoEndpoint {
  clickTrackingParams: string;
  watchEndpoint: WatchEndpoint;
}

export interface WatchEndpoint {
  videoId: string;
  playlistId: string;
  index: number;
  playlistSetVideoId: string;
  loggingContext: LoggingContext;
}

export interface LoggingContext {
  vssLoggingContext: VssLoggingContext;
}

export interface VssLoggingContext {
  serializedContextData: string;
}

export interface PlayerOverlays {
  playerOverlayRenderer: PlayerOverlayRenderer;
}

export interface PlayerOverlayRenderer {
  actions: PlayerOverlayRendererAction[];
  browserMediaSession: BrowserMediaSession;
}

export interface PlayerOverlayRendererAction {
  likeButtonRenderer: LikeButtonRenderer;
}
// ...
export interface BrowserMediaSession {
  browserMediaSessionRenderer: BrowserMediaSessionRenderer;
}

export interface BrowserMediaSessionRenderer {
  album: Title;
  thumbnailDetails: ThumbnailDetails;
}

export interface Title {
  runs: TitleRun[];
}

export interface TitleRun {
  text: string;
}

export interface ThumbnailDetails {
  thumbnails: Thumbnail[];
}

export interface Thumbnail {
  url: string;
  width: number;
  height: number;
}

export interface ResponseContext {
  serviceTrackingParams: ServiceTrackingParam[];
}

export interface ServiceTrackingParam {
  service: string;
  params: Param[];
}

export interface Param {
  key: string;
  value: string;
}

export interface VideoReporting {
  reportFormModalRenderer: ReportFormModalRenderer;
}

export interface ReportFormModalRenderer {
  optionsSupportedRenderers: OptionsSupportedRenderers;
  trackingParams: string;
  title: Title;
  submitButton: Button;
  cancelButton: Button;
  footer: Footer;
}
// ...
export type VideoDataChangeValue = Record<string, unknown> & {
  videoId: string;
  title: string;
  author: string;

  playlistId: string;
  isUpcoming: boolean;
  loading: boolean;

  lengthSeconds: number;

  /**
   * Internal variable (for album data)
   **/
  ytmdWatchNextResponse?: WatchNextResponse;
};